requirements.txt
azure-functions
opencv-python
numpy
Pillow